feat: work on todos
