--- package/src/sitemap/virtual.d.ts
@@ -1,5 +1,0 @@
-declare module "i18n:astro/sitemap" {
-	const sitemap: (args: import("./route-config.js").CallbackSchema) => void;
-
-	export default sitemap;
-}
