+++ package/src/stubs/sitemap.d.ts
@@ -1,0 +1,5 @@
+declare module "i18n:astro/sitemap" {
+	const sitemap: (args: import("../sitemap/route-config.js").CallbackSchema) => void;
+
+	export default sitemap;
+}
